Rename `react-components/` to `react-component`
