+++ addon/components/react-component/yield-wrapper.js
